refactored interface matrix
